QCamera2: Port HAL for 64 bit platforms
- Added -Wextra compiler warning level.
- Avoiding implicit datatype conversions
  and also pointer to integer conversions,
  including literals.
- Avoiding pointer conversions between different types
  other than (void*).
- Using 'size_t' to store memory size and indexes.
- Removed some unused variables.
- Changed datatype of some iterators to match their
  corresponding limiter types.
- Using double instead of float where possible.
- Changed bitmasks to unsigned types.
- Updated printf/scanf format string specifiers to match
  actual type of arguments.
- Using 'String8' instead of 'char[]' where possible
  at QCameraParameters.
- Checking result of QCameraMemory::getSize().
- Fixes build errors and warnings.

CRs-Fixed: 641033

+++ QCamera2/HAL/QCameraPostProc.h
@@ -146,7 +146,7 @@ private:
 
     QCameraStreamMemory *      m_pJpegOutputMem;
     QCameraExif *              m_pJpegExifObj;
-    int8_t                     m_bThumbnailNeeded;
+    uint32_t                   m_bThumbnailNeeded;
     QCameraReprocessChannel *  m_pReprocChannel;
 
     QCameraQueue m_inputPPQ;            // input queue for postproc
